libcrunch_cil_inlines.h: update to silence warnings caused by cache refactoring in liballocs.

+++ include/libcrunch_cil_inlines.h
@@ -578,11 +578,11 @@ extern inline int (__attribute__((always_inline,gnu_inline,used)) __can_hold_poi
 		// FIXME: replace these nasty hardcoded constants with
 		// use of macros that we can generate from dwarfidl of struct uniqtype
 		// (for clients that can't afford to include uniqtype-defs.h)
-		char kind = *((char*) hit->t + 0xc);
+		char kind = *((char*)(unsigned long) hit->t + 0xc);
 		if ((kind & 0xf) == 8)
 		{
 			char related[8];
-			__builtin_memcpy(related, (char*) hit->t + 0x18, 8);
+			__builtin_memcpy(related, (char*)(unsigned long) hit->t + 0x18, 8);
 			struct __liballocs_memrange_cache_entry_s *hit2 = __liballocs_memrange_cache_lookup(
 				&__liballocs_ool_cache,
 				value, *(void**) related, 0);
